dicom/dicom_service/requirements.txt
pynetdicom == 2.0.2
requests == 2.31.0
flask == 3.0.0
pytest == 7.4.3
pytest-cov == 4.1.0
APScheduler == 3.10.4
Flask-JWT-Extended == 4.6.0
PyYAML==6.0.1